+++ like-scene/src/index.ts
@@ -28,9 +28,9 @@
  * @module scene
  */
 
-import { likeDispatch } from '@like2d/like';
-import type { LikeEvent } from '@like2d/like';
-import type { Like, LikeHandlers } from '@like2d/like';
+import { likeDispatch } from 'like2d';
+import type { LikeEvent } from 'like2d';
+import type { Like, LikeHandlers } from 'like2d';
 
 /** {@include creating-scenes.md} */
 export type Scene = (like: Like, scenes: SceneManager) => SceneInstance;
@@ -180,7 +180,7 @@ export class SceneManager {
         top.instance = this.instantiate(top.factory);
       }
     } else {
-      likeDispatch(this.like, { type: 'load', args: [], timestamp: this.like.timer.getTime() })
+      likeDispatch(this.like, { type: 'load', args: [] })
     }
     return oldTop?.factory;
   }
@@ -190,7 +190,7 @@ export class SceneManager {
    */
   instantiate<T extends SceneEx<SceneInstance>>(scene: T): InstantiateReturn<T> {
     const inst = scene(this.like, this);
-    likeDispatch(inst, { type: 'load', args: [], timestamp: this.like.timer.getTime() });
+    likeDispatch(inst, { type: 'load', args: [] });
     return inst as InstantiateReturn<T>;
   }
 
@@ -203,7 +203,7 @@ export class SceneManager {
     const scene = this.scenes.at(pos);
     if (scene) {
       if (scene.instance) {
-        likeDispatch(scene.instance, { type: 'quit', args: [], timestamp: this.like.timer.getTime() })
+        likeDispatch(scene.instance, { type: 'quit', args: [] })
       }
       scene.instance = undefined;
     }